Fixed the issue with not seeing titles.
